refactor update & delete category

+++ controllers/categories-controller.js
@@ -28,6 +28,20 @@ const addCategory = async (req, res) => {
 
 const deleteCategory = async (req, res) => {
   try {
+    const cat = await Category.findOne({ _id: req.params.id });
+
+    if (!cat) {
+      return res.status(400).json({ msg: "category not found!" });
+    }
+    const linkedProduct = await Product.find({ category: cat.name });
+
+    //check if any product has same category before delete
+    if (linkedProduct.length !== 0) {
+      return res.status(400).json({
+        msg: "can't delete this category because there are products associated with it !",
+      });
+    }
+
     const data = await Category.deleteOne({ _id: req.params.id });
     res.status(200).json(data);
   } catch (error) {
@@ -37,6 +51,15 @@ const deleteCategory = async (req, res) => {
 
 const editCategory = async (req, res) => {
   try {
+    const cat = await Category.find({ _id: req.params.id });
+
+    //update product with same category
+    await Product.updateMany(
+      { category: cat[0].name },
+      { $set: { category: req.body.name } }
+    );
+
+    //update category
     const data = await Category.findByIdAndUpdate(
       req.params.id,
       {
@@ -44,6 +67,7 @@ const editCategory = async (req, res) => {
       },
       { new: true }
     );
+
     res.status(200).json(data);
   } catch (error) {
     res.status(400).json({ msg: error.message });
